// src/app/models/empresa.model.ts
export interface Empresa{
    empid?: number;
    empnit?: string;
    empnombre?: string;
    emprepresentante?: string;
    empdepartamento?: string;
    empciudad?: string;  
    empdireccion?: string;
    emptelefono?: string;
    empactiva?: string;
    empfechaini?: Date;
    emparea?:string;
}